Controlador de Materiales
Este commit fue forzado para poder arreglar conflictos de versiones

+++ ProyectoFinal/ProyectoFinal/Modelo/Produccion/Materiales.cs
@@ -11,7 +11,7 @@ namespace ProyectoFinal.Modelo.Produccion
         public int IdMaterial { get; set; }
         public String Descripcion { get; set; }
         public double Costo { get; set; }
-        public int Existencia { get; set; }
+        public double Existencia { get; set; }
         public TiposMateriales Tipo { get; set; }
 
         public enum TiposMateriales : int
